requirements.txt
streamlit
torch
torchvision
opencv-python
numpy
networkx
matplotlib
pandas
google-generativeai
